+++ knock/src/components/Modal/CrateQuestion.module.scss
@@ -1,5 +1,17 @@
 @import '../../core/styles/global.scss';
 
+.modal-container {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: rgba(0, 0, 0, 0.5);
+}
+
 .modal-open-button {
   display: flex;
   width: 13rem;
@@ -17,6 +29,13 @@
 }
 
 .modal-content {
+  width: 38.25rem;
+  height: 28.375rem;
+  flex-shrink: 0;
+  border-radius: 1.5rem;
+  @include Shadow(3);
+  background-color: $Grayscale-10;
+
   &__header-question {
     display: flex;
     width: 33.25rem;
@@ -78,8 +97,12 @@
     margin: auto;
   }
 
-  &__question-button {
+  &__question-button-container {
     text-align: center;
+  }
+
+  &__question-button {
+    width: 33.25rem;
   }
 
   &__container {
