+++ knock/src/components/Modal/Error.module.scss
@@ -1,9 +1,21 @@
 @import '../../core/styles/global.scss';
 
 .error-modal {
-  &__content {
+  &__container {
     width: 100%;
     height: 100%;
+    position: fixed;
+    top: 0;
+    left: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: rgba(0, 0, 0, 0.5);
+  }
+
+  &__content {
+    width: 19.25rem;
+    height: 14.375rem;
     flex-shrink: 0;
     border-radius: 1.5rem;
     @include Shadow(3);
