+++ knock/src/components/Modal/crateQuestion.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import style from './CrateQuestion.module.scss';
 import UButton from '../../core/ui/buttons/UButton/UButton';
 import Image from '../../core/ui/CommonImage/Image';
@@ -6,7 +5,7 @@ import Textarea from '../../core/ui/Textarea/Textarea';
 import Icon from '../../core/ui/CommonIcon/icon';
 import messageicon from '../../core/assets/icon/Messages.svg';
 import closeicon from '../../core/assets/icon/Close.svg';
-import CommonModal from '../../core/ui/Modal/Modal';
+import CommonModal from '../../core/ui/Modal/CommonModal';
 
 interface CrateQuestionProps {
   name: string;
@@ -18,6 +17,7 @@ interface CrateQuestionProps {
   onChange: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
   isActivation: boolean;
   isLoading: boolean;
+  textareaValue: string;
 }
 
 const CrateQuestion: React.FC<CrateQuestionProps> = ({
@@ -30,76 +30,64 @@ const CrateQuestion: React.FC<CrateQuestionProps> = ({
   onChange,
   isActivation,
   isLoading,
+  textareaValue,
 }) => {
-  const [textareaValue, setTextareaValue] = useState('');
-
   const handleSendClick = () => {
     onSubmit(textareaValue);
-    setTextareaValue('');
-  };
-
-  const handleTextareaChange = (
-    event: React.ChangeEvent<HTMLTextAreaElement>,
-  ) => {
-    setTextareaValue(event.target.value);
-    onChange(event);
   };
 
   return (
-    <>
-      <CommonModal isOpen={isOpen} onClose={onClose}>
-        <div className={style['modal-content']}>
-          <div className={style['modal-content__header-question']}>
-            <div className={style['modal-content__question']}>
-              <Icon
-                className={style['modal-content__message-icon']}
-                src={messageicon}
-                alt="메시지 아이콘"
-              />
-              <h1 className={style['modal-content__title']}>
-                질문을 작성하세요
-              </h1>
-            </div>
-            <button className={style['modal-content__close-icon']}>
-              <img onClick={onClose} src={closeicon} alt="닫기 아이콘" />
-            </button>
+    <CommonModal isOpen={isOpen} onClose={onClose}>
+      <div className="modal-content">
+        <div className={style['modal-content__header-question']}>
+          <div className={style['modal-content__question']}>
+            <Icon
+              className={style['modal-content__message-icon']}
+              src={messageicon}
+              alt="메시지 아이콘"
+            />
+            <h1 className={style['modal-content__title']}>질문을 작성하세요</h1>
           </div>
-          <label htmlFor={style['textareaContent']}>
-            <div className={style['modal-content__profile-container']}>
-              <span className={style['modal-content__to-label']}>To.</span>
-              <Image
-                imageClassName={style['modal-content__profile-image']}
-                src={src}
-                alt={alt}
-              />
-              {name}
-            </div>
-          </label>
+          <button className={style['modal-content__close-icon']}>
+            <img onClick={onClose} src={closeicon} alt="닫기 아이콘" />
+          </button>
+        </div>
+        <label htmlFor={style['textareaContent']}>
+          <div className={style['modal-content__profile-container']}>
+            <span className={style['modal-content__to-label']}>To.</span>
+            <Image
+              imageClassName={style['modal-content__profile-image']}
+              src={src}
+              alt={alt}
+            />
+            {name}
+          </div>
+        </label>
 
-          <div className={style['modal-content__container']}>
-            <div className={style['modal-content__textarea']}>
-              <Textarea
-                placeholder="질문을 입력하세요"
-                id="textareaContent"
-                name={name}
-                value={textareaValue}
-                onChange={handleTextareaChange}
-              />
-            </div>
-            <div className={style['modal-content__question-button']}>
-              <UButton
-                handleClick={handleSendClick}
-                type="box"
-                isLightTheme={false}
-                isDisalbed={isActivation || isLoading}
-              >
-                질문 보내기
-              </UButton>
-            </div>
+        <div className={style['modal-content__container']}>
+          <div className={style['modal-content__textarea']}>
+            <Textarea
+              placeholder="질문을 입력하세요"
+              id="textareaContent"
+              name={name}
+              value={textareaValue}
+              onChange={onChange}
+            />
+          </div>
+          <div className={style['modal-content__question-button-container']}>
+            <UButton
+              className={style['modal-content__question-button']}
+              handleClick={handleSendClick}
+              type="box"
+              isLightTheme={false}
+              isDisalbed={isActivation || isLoading}
+            >
+              질문 보내기
+            </UButton>
           </div>
         </div>
-      </CommonModal>
-    </>
+      </div>
+    </CommonModal>
   );
 };
